backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.3.0
sentence-transformers==2.2.2
spacy==3.7.2
numpy==1.24.3
torch==2.0.1
transformers==4.35.0
# Authentication and security
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.7

# Additional ML and data processing
pandas==2.0.3
tensorflow==2.13.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
wordcloud==1.9.2
Pillow==10.0.1
reportlab==4.0.4
joblib==1.3.2
scipy==1.11.4